+++ lmcache/cli/commands/bench/test_cache.py
